Fixes for project properties / meta tags
- The "subtitle" tag was missing
- A few other tags like "workNumber", "movementTitle", "movementNumber" were lost because they were treated as "standard tags" but not stored in `ProjectMeta`.
