+++ src/appshell/internal/istartupscenario.h
@@ -24,7 +24,7 @@
 
 #include "modularity/imoduleinterface.h"
 
-#include "project/projecttypes.h"
+#include "project/types/projecttypes.h"
 
 namespace mu::appshell {
 class IStartupScenario : MODULE_EXPORT_INTERFACE
